+++ mType/vm/compiler/emission/BytecodeEmitter.hpp
@@ -20,10 +20,10 @@ namespace vm::compiler::emission
 
         // Bytecode emission with source location tracking
         void emitWithLocation(bytecode::OpCode opcode, ast::ASTNode* node);
-        void emitWithLocation(bytecode::OpCode opcode, uint32_t operand, ast::ASTNode* node);
-        void emitWithLocation(bytecode::OpCode opcode, uint32_t operand1, uint32_t operand2, ast::ASTNode* node);
-        void emitWithLocation(bytecode::OpCode opcode, uint32_t operand1, uint32_t operand2, uint32_t operand3, ast::ASTNode* node);
-        void emitWithLocation(bytecode::OpCode opcode, const std::vector<uint32_t>& operands, ast::ASTNode* node);
+        void emitWithLocation(bytecode::OpCode opcode, uint64_t operand, ast::ASTNode* node);
+        void emitWithLocation(bytecode::OpCode opcode, uint64_t operand1, uint64_t operand2, ast::ASTNode* node);
+        void emitWithLocation(bytecode::OpCode opcode, uint64_t operand1, uint64_t operand2, uint64_t operand3, ast::ASTNode* node);
+        void emitWithLocation(bytecode::OpCode opcode, const std::vector<uint64_t>& operands, ast::ASTNode* node);
 
         // Jump management
         size_t emitJump(bytecode::OpCode jumpOp, ast::ASTNode* node = nullptr);
